Crate stdbench

This library contains all necessary tools to run a PISA benchmark on a collection of a significant size.

Modules

build
Anything that has to do with building either forward or inverted index, including parsing and compressing.

config
Experiment configuration, which is used throughout a run, and mostly defined in an external YAML configuration file (with several exceptions).

error
Defines the error type and its formatting and conversion.

executor
Objects and functions dealing with executing PISA command line tools.

source
Objects and functions dealing with retrieving and preparing PISA executables.

Macros

execute
Executes a $cmd, checks for results, and returns an error with $errmsg message. It is designed to be similar to ? operator, removing bulky boilerplate from functions that execute many consecutive commands.

Enums

Stage
Available stages of the experiment.

Functions

ensure_parent_exists
If the parent directory of path does not exist, create it.

printed
Prints the passed command and returns it back.